+++ src/main/java/fr/davipro/chatopbackend/model/User.java
@@ -15,11 +15,11 @@ import jakarta.persistence.Table;
 @Table(name = "users")
 public class User {
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -65,7 +65,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String email;
